+++ tests/test_theme.py
@@ -1,0 +1,30 @@
+#  Copyright 2022 Feedzai
+#
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+"""Test suite for theme registration and enablement."""
+
+import altair as alt
+
+
+def test_theme_is_available() -> None:
+    """Check if the theme was registered via an entry point and is available as an Altair theme."""
+    assert "feedzai" in alt.theme.names()
+
+
+def test_theme_is_enabled() -> None:
+    """Check if the theme is not active by default and can be enabled."""
+    assert alt.theme.active != "feedzai"
+
+    alt.theme.enable("feedzai")
+    assert alt.theme.active == "feedzai"